requirements.txt
paho-mqtt==1.6.1
Unidecode==1.3.6
